moved tests to transport for doppler_factor
